core: Major internal refactoring to simplify incremental compilation   (lineinfo managed in tokenlist).
The lineinfo attributes in the Metalua 0.5 git branch are difficult to traverse
and update.  Under this change, the lineinfo information is extracted into
a ordered linear token list that cross-references back the the AST.
The lineinfo in the AST is no longer used nor updated.  This change
simplifies the code and is less error prone.