analysis - lookup function's static info from its dynamic debug info.
Example:

  -- lib.lua
  foo = {}
  function foo.bar() return {x=1} end

  -- test.lua
  require 'lib'
  local xx = foo.bar()

Here, foo.bar is resolved to a function value by dynamic analysis.
The debug info on that function value is queried to obtain source/line
position of the function definition.  The function definition was
previously analyzed by static analysis (inferring return values) and
stored in static info, and we now look that up.
Therefore, we can now infer that xx is a table.